ORder tests and models bigint changed to integer

+++ migrations/sqls/20210307122013-orders-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
   id SERIAL PRIMARY KEY, 
   completed BOOLEAN, 
-  user_id BIGINT REFERENCES users(id)
+  user_id INTEGER REFERENCES users(id)
   );
